requirements.txt
anthropic==0.40.0
langchain==0.3.14
langgraph==0.2.60
langchain-anthropic==0.3.6
chromadb==0.5.23
sentence-transformers==3.3.1
streamlit==1.40.2
networkx==3.4.2
matplotlib==3.9.3
plotly==5.24.1
google-search-results==2.4.2
pytrends==4.9.2
textstat==0.7.4
python-docx==1.1.2
PyMuPDF==1.24.14
pyyaml==6.0.2
rich==13.9.4